fix:utils関連ファイルの整理
